requirements-kiro.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
anthropic==0.25.8
google-generativeai==0.4.0
boto3==1.34.0
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.8
langchain-anthropic==0.1.0
langchain-google-genai==0.0.10
langchain-community==0.0.10
pydantic-ai==0.0.1
structured-logger==0.5.0
aiohttp==3.9.1
aiofiles==23.2.1
tqdm==4.66.1
tenacity==8.2.3
pyyaml==6.0